Move package "client/connection" to "core/connection"
