+++ core/connection/interface.go
