+++ core/connection/manager.go
