+++ core/connection/manager_test.go
