+++ core/node/node.go
@@ -25,10 +25,10 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/julienschmidt/httprouter"
 	"github.com/mysterium/node/blockchain"
-	"github.com/mysterium/node/client/connection"
 	"github.com/mysterium/node/communication"
 	nats_dialog "github.com/mysterium/node/communication/nats/dialog"
 	nats_discovery "github.com/mysterium/node/communication/nats/discovery"
+	"github.com/mysterium/node/core/connection"
 	"github.com/mysterium/node/identity"
 	"github.com/mysterium/node/identity/registry"
 	"github.com/mysterium/node/ip"
