+++ core/promise/promise_test.go
