+++ tequilapi/endpoints/connection.go
@@ -23,7 +23,7 @@ import (
 
 	log "github.com/cihub/seelog"
 	"github.com/julienschmidt/httprouter"
-	"github.com/mysterium/node/client/connection"
+	"github.com/mysterium/node/core/connection"
 	"github.com/mysterium/node/identity"
 	"github.com/mysterium/node/ip"
 	"github.com/mysterium/node/openvpn/middlewares/client/bytescount"
