+++ tequilapi/endpoints/location.go
@@ -21,7 +21,7 @@ import (
 	"net/http"
 
 	"github.com/julienschmidt/httprouter"
-	"github.com/mysterium/node/client/connection"
+	"github.com/mysterium/node/core/connection"
 	"github.com/mysterium/node/location"
 	"github.com/mysterium/node/tequilapi/utils"
 )
